rag-stress-test-arena/requirements.txt
requests
psycopg2-binary
redis[hiredis]
numpy
matplotlib